Refactor javascript.
* Push each functional area to its own file. Thus sufia.js becomes just a
manifest.
* Leverage bootstrap's alert function instead of having our own.
* Reenabled feature spec for proxy deposit
