+++ app/assets/javascripts/sufia/select_submit.js
@@ -1,0 +1,12 @@
+//override the blacklight default to submit
+//form when sort by or show per page change
+Blacklight.do_select_submit = function() {
+  $(Blacklight.do_select_submit.selector).each(function() {
+        var select = $(this);
+        select.closest("form").find("input[type=submit]").show();
+        select.bind("change", function() {
+          return false;
+        });
+    });
+};
+Blacklight.do_select_submit.selector = "form.sort select, form.per_page select";
