+++ app/search_builders/sufia/my_shares_search_builder.rb
@@ -2,5 +2,5 @@
 class Sufia::MySharesSearchBuilder < Sufia::SearchBuilder
   include Sufia::MySearchBuilderBehavior
 
-  self.default_processor_chain = default_processor_chain - [:filter_models] + [:show_only_shared_files, :show_only_file_sets]
+  self.default_processor_chain = default_processor_chain - [:filter_models] + [:show_only_shared_files]
 end
