+++ app/search_builders/sufia/search_builder.rb
@@ -18,13 +18,6 @@ class Sufia::SearchBuilder < Blacklight::SearchBuilder
     ]
   end
 
-  def show_only_file_sets(solr_parameters)
-    solr_parameters[:fq] ||= []
-    solr_parameters[:fq] += [
-      ActiveFedora::SolrQueryBuilder.construct_query_for_rel(has_model: ::FileSet.to_class_uri)
-    ]
-  end
-
   def show_only_generic_works(solr_parameters)
     solr_parameters[:fq] ||= []
     solr_parameters[:fq] += [
