+++ app/views/curation_concerns/base/_browse_everything.html.erb
@@ -1,5 +1,5 @@
 <div class="alert alert-success">
-  <%=  t("sufia.upload.cloud_timeout_message", contact_href: link_to(t("sufia.upload.alert.contact_href_text"), sufia.contact_form_index_path)) %>
+  <%= t("sufia.upload.cloud_timeout_message", contact_href: link_to(t("sufia.upload.alert.contact_href_text"), sufia.contact_form_index_path)) %>
 </div>
 <%= form_tag(sufia.uploads_path, id: 'browse_everything_form', method: 'post') do %>
     <%= hidden_field_tag(:upload_set_id, @upload_set_id) %>
