+++ app/views/curation_concerns/base/_permission_form.html.erb
@@ -8,12 +8,10 @@
 </h2>
 
 <div class="alert alert-info hidden" id="save_perm_note">Permissions are <strong>not</strong> saved until the &quot;Save&quot; button is pressed at the bottom of the page.</div>
-<div class="alert alert-warning hidden" id="permissions_error">
-  <a class="close" data-dismiss="alert" href="#">
-    <span class="sr-only">Close this alert</span>
-    <span aria-hidden="true">×</span>
-  </a>
-<span id="permissions_error_text"></span></div>
+<div class="alert alert-warning hidden alert-dismissable" role="alert" id="permissions_error">
+  <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+  <span id="permissions_error_text"></span>
+</div>
 
 <%= render 'curation_concerns/base/form_permission', f: f %>
